fix typo

+++ cpp/src/arrow/array/util.cc
@@ -166,9 +166,9 @@ class ArrayDataEndianSwapper {
   }
 
   Status Visit(const DayTimeIntervalType& type) {
-    auto data = reinterpret_cast<const uint64_t*>(data_->buffers[1]->data());
+    auto data = reinterpret_cast<const uint32_t*>(data_->buffers[1]->data());
     ARROW_ASSIGN_OR_RAISE(auto new_buffer, AllocateBuffer(data_->buffers[1]->size()));
-    auto new_data = reinterpret_cast<uint64_t*>(new_buffer->mutable_data());
+    auto new_data = reinterpret_cast<uint32_t*>(new_buffer->mutable_data());
     int64_t length = length_;
     for (int64_t i = 0; i < length; i++) {
       auto idx = i * 2;
